+++ src/lib.rs
@@ -1,16 +1,9 @@
 mod baker;
 mod bits;
 pub mod circuit;
-mod buffer;
-pub mod bytecode;
-mod constants;
 mod interpreter;
-mod number;
 mod proof;
-mod rng;
 mod simulation;
-mod simulation2;
 
 pub use proof::{prove, verify, Proof};
-pub use circuit::{Circuit, Instruction};
-pub use simulation2::Error;
+pub use simulation::Error;
